ファイルを切り分ける
