+++ src/features/bar/About/index.tsx
@@ -1,0 +1,132 @@
+"use client";
+
+import React, { useEffect, useRef } from "react";
+import Image from "next/image";
+import styles from "./style.module.scss";
+
+type BarAboutProps = {
+  onIntersect: (isIntersected: boolean) => void;
+};
+
+const BarAbout: React.FC<BarAboutProps> = ({ onIntersect }) => {
+  const targetRef = useRef<HTMLDivElement | null>(null);
+
+  useEffect(() => {
+    if (!targetRef.current) return;
+
+    const observer = new IntersectionObserver(
+      ([entry]) => {
+        if (entry.isIntersecting) {
+          onIntersect(true);
+        } else {
+          onIntersect(false);
+        }
+      },
+      {
+        threshold: 0,
+        rootMargin: "0px 0px -92% 0px", // 画面上部に到達したらトリガー
+      }
+    );
+
+    observer.observe(targetRef.current);
+  }, [onIntersect]);
+
+  return (
+    <>
+      <div ref={targetRef} id="#BarAbout" className={styles.BarAbout}>
+        <div className={styles.BarAbout__texts}>
+          <h2 className={styles.BarAbout__title}>
+            <p className={styles.BarAbout__titleJp}>わととについて</p>
+            <p className={styles.BarAbout__titleEn}>about watoto</p>
+          </h2>
+          <div className={styles.BarAbout__content}>
+            <p className={styles.BarAbout__contentJp}>
+              {/* あとでJson化したい */}
+              京都下鴨本通沿いに位置する「わとと」は、
+              <br />
+              訪れる人々に焦点を当てた飲食店です。
+              <br />
+              <br />
+              私たちは、人と人とのつながり、
+              <br />
+              そこで生まれる時間を大切にしています。
+              <br />
+              <br />
+              わととではお酒とコーヒー、
+              <br />
+              体をいたわる薬膳食や発酵食品を提供しています。
+              <br />
+              <br />
+              しかし、
+              <br />
+              ある時には音楽のイベントが開かれ、
+              <br />
+              またある時には習字教室が開かれたり、
+              <br />
+              その活動は一つの言葉では表せません。
+              <br />
+              <br />
+              既定の形にとらわれず、
+              <br />
+              水のように柔軟に形を変え、
+              <br />
+              さまざまな人々の過ごし方に
+              <br />
+              寄り添っていきたいと考えています。
+            </p>
+            <p className={styles.BarAbout__contentEn}>
+              Located along Shimogamo-Hondori in Kyoto,
+              <br />
+              watoto is an establishment that focuses on
+              <br />
+              the people who visit us.
+              <br />
+              <br />
+              We cherish human connections and
+              <br />
+              the moments that arise from them.
+              <br />
+              <br />
+              While watoto serves alcohol, coffee,
+              <br />
+              nourishing Chinese herbal cuisine,
+              <br />
+              and fermented food.
+              <br />
+              <br />
+              But our activities cannot be defined by
+              <br />a single description -
+              <br />
+              sometimes we host music events,
+              <br />
+              other times we hold calligraphy classes.
+              <br />
+              <br />
+              Like water,
+              <br />
+              we aim to remain flexible and adapt
+              <br />
+              our form without being bound by
+              <br />
+              conventional structures,
+              <br />
+              accommodating the diverse ways in
+              <br />
+              which people choose to spend their time.
+            </p>
+          </div>
+        </div>
+
+        <Image
+          src="/bar/about/1.JPG"
+          alt="About バー店内の写真"
+          width={720}
+          height={967}
+          className={styles.BarAbout__image}
+        />
+      </div>
+    </>
+  );
+};
+
+export default BarAbout;
