+++ src/features/bar/About/style.module.scss
@@ -1,0 +1,69 @@
+@use "../../../styles/mixin" as m;
+
+// About
+.BarAbout {
+  position: relative;
+  z-index: var(--z-index-elevation-1);
+  margin-top: -48px;
+  padding: 180px 0 100px 100px;
+  display: flex;
+  justify-content: space-between;
+  gap: 60px;
+  background-color: var(--color-base-background);
+  border-radius: 60px 60px 0 0;
+
+  .BarAbout__texts {
+    padding-top: 240px;
+    width: 100%;
+  }
+
+  .BarAbout__title {
+    padding-bottom: 20px;
+    color: var(--color-black);
+    border-bottom: 1px solid var(--color-border-gray);
+  }
+
+  .BarAbout__titleJp {
+    font-size: 18px;
+  }
+
+  .BarAbout__titleEn {
+    margin-top: 4px;
+    color: var(--color-text-gray-2);
+    font-size: 15px;
+    line-height: 1;
+  }
+
+  .BarAbout__content {
+    margin-top: 20px;
+    display: flex;
+    gap: 60px;
+    justify-content: start;
+  }
+
+  .BarAbout__contentJp {
+    min-width: 306px;
+    font-family: var(--font-family-text-jp);
+    font-size: 13px;
+    color: var(--color-black);
+    line-height: 2;
+    letter-spacing: var(--letter-space-text-jp);
+  }
+
+  .BarAbout__contentEn {
+    min-width: 306px;
+    font-family: var(--font-family-text-en);
+    font-size: 15px;
+    color: var(--color-text-gray);
+    line-height: 1.4;
+    letter-spacing: var(--letter-space-text-en);
+  }
+
+  .BarAbout__image {
+    width: 30%;
+    max-height: 720px;
+    object-fit: cover;
+    aspect-ratio: 800 / 1600;
+    border-radius: 40px 0 0 40px;
+  }
+}
