+++ src/features/bar/Concepts/index.tsx
@@ -1,0 +1,8 @@
+"use client";
+
+import React from "react";
+import style from "./style.module.scss";
+
+const BarConcepts: React.FC = ({}) => <div className={style.BarConcepts}></div>;
+
+export default BarConcepts;
