+++ src/features/bar/Concepts/style.module.scss
@@ -1,0 +1,4 @@
+@use "../../../styles/mixin" as m;
+
+.Concepts {
+}
