+++ src/features/bar/Top/style.module.scss
@@ -1,6 +1,6 @@
-@use "../../styles/mixin" as m;
+@use "../../../styles/mixin" as m;
 
-.Bar {
+.BarTop {
   $padding-frame-vertical: 40px;
   $padding-frame-horizontal: 48px;
 
@@ -11,24 +11,12 @@
     padding: 16px;
   }
 
-  .Bar__topVideo {
-    position: fixed;
-    top: 50%;
-    left: 50%;
-    z-index: var(--z-index-elevation-2);
-    width: 230px;
-    height: 320px;
-    object-fit: cover;
-    border-radius: 2px;
-    transform: translate(-50%, -50%);
-  }
-
-  .Bar__slider {
+  .BarTop__slider {
     height: 100svh;
     overflow: hidden;
   }
 
-  .Bar__image {
+  .BarTop__image {
     width: 100%;
     height: 100%;
     object-fit: cover;
@@ -41,7 +29,7 @@
     }
   }
 
-  .Bar__text {
+  .BarTop__text {
     position: relative;
     bottom: 148px;
     left: $padding-frame-vertical;
@@ -62,28 +50,20 @@
       letter-spacing: 0.2em;
     }
 
-    .Bar__textJp {
+    .BarTop__textJp {
       font-size: 13px;
       line-height: 1.6;
     }
-    .Bar__textEn {
+    .BarTop__textEn {
       margin-top: 8px;
       font-size: 9px;
       line-height: 1.4;
     }
   }
 
-  .Bar__scrollIndicator {
+  .BarTop__scrollIndicator {
     position: absolute;
     right: -12px;
     bottom: 68px;
-  }
-
-  .Bar__imageSecond {
-    position: sticky;
-    top: 0px;
-    margin-top: -96px;
-    width: 100%;
-    height: 100vh;
   }
 }
